requirements.txt
streamlit
pdfplumber 
pandas 
requests 
python-dotenv